RecStrip 2.6: - weiterer Bugfix in PESProcessor: Remaining Bytes griff auf Data[184] zu - Typ-Konvertierungen für x64

+++ CutProcessor.c
@@ -216,7 +216,7 @@ static bool CutFileDecodeTxt(FILE *fCut, unsigned long long *OutSavedSize)
       if(c) *c = '\0';
 
       // Remove line breaks in the end
-      p = strlen(Buffer);
+      p = (int)strlen(Buffer);
       while (p && (Buffer[p-1] == '\r' || Buffer[p-1] == '\n' || Buffer[p-1] == ';'))
         Buffer[--p] = '\0';
